Action cable progress

+++ db/migrate/20190918185506_create_groups.rb
@@ -2,7 +2,7 @@ class CreateGroups < ActiveRecord::Migration[5.2]
   def change
     create_table :groups do |t|
       t.string :group_code
-      t.integer :bill_total
+      t.float :bill_total
       t.timestamps
     end
   end
